add magic method to pipetterank and move to helpers file
